+++ crates/sncast/src/helpers/commands.rs
